Moved realmEmojiReducers to realmReducer.
Fix:#607
