+++ src/realm/__tests__/realmtReducers-test.js
@@ -2,7 +2,8 @@ import realmReducers from '../realmReducers';
 import {
   ACCOUNT_SWITCH,
   SAVE_TOKEN_GCM,
-  DELETE_TOKEN_GCM
+  DELETE_TOKEN_GCM,
+  EVENT_REALM_EMOJI_UPDATE
 } from '../../actionConstants';
 
 describe('realmReducers', () => {
@@ -15,6 +16,7 @@ describe('realmReducers', () => {
       const expectedState = {
         twentyFourHourTime: false,
         'gcmToken': '',
+        emoji: {},
       };
 
       const actualState = realmReducers(initialState, action);
@@ -28,6 +30,7 @@ describe('realmReducers', () => {
       const initialState = {
         twentyFourHourTime: false,
         'gcmToken': '',
+        emoji: { customEmoji1: {} },
       };
       const action = {
         type: SAVE_TOKEN_GCM,
@@ -36,6 +39,7 @@ describe('realmReducers', () => {
       const expectedState = {
         twentyFourHourTime: false,
         'gcmToken': 'new-key',
+        emoji: { customEmoji1: {} },
       };
 
       const actualState = realmReducers(initialState, action);
@@ -50,6 +54,7 @@ describe('realmReducers', () => {
       const initialState = {
         twentyFourHourTime: false,
         'gcmToken': 'old-key',
+        emoji: { customEmoji1: {} },
       };
       const action = {
         type: DELETE_TOKEN_GCM,
@@ -57,11 +62,44 @@ describe('realmReducers', () => {
       const expectedState = {
         twentyFourHourTime: false,
         'gcmToken': '',
+        emoji: { customEmoji1: {} },
       };
 
       const actualState = realmReducers(initialState, action);
 
       expect(actualState).toEqual(expectedState);
+    });
+  });
+
+  describe('EVENT_REALM_EMOJI_UPDATE', () => {
+    test('update state to new realm_emoji', () => {
+      const prevState = {
+        twentyFourHourTime: false,
+        'gcmToken': 'key',
+        emoji: {},
+      };
+      const action = {
+        eventId: 4,
+        id: 4,
+        op: 'update',
+        realm_emoji: {
+          customEmoji1: {},
+          customEmoji2: {},
+        },
+        type: EVENT_REALM_EMOJI_UPDATE
+      };
+      const expectedState = {
+        twentyFourHourTime: false,
+        'gcmToken': 'key',
+        emoji: {
+          customEmoji1: {},
+          customEmoji2: {},
+        }
+      };
+
+      const newState = realmReducers(prevState, action);
+
+      expect(newState).toEqual(expectedState);
     });
   });
 });
